Move `SearchSpecification`, `ConditionsTreeNode` and `LocalSearch` to `:legacy:search` module
