+++ feature/widget/message-list/src/main/kotlin/app/k9mail/feature/widget/message/list/MessageListConfig.kt
@@ -1,7 +1,7 @@
 package app.k9mail.feature.widget.message.list
 
 import app.k9mail.legacy.account.Account.SortType
-import com.fsck.k9.search.LocalSearch
+import app.k9mail.legacy.search.LocalSearch
 
 internal data class MessageListConfig(
     val search: LocalSearch,
