+++ legacy/core/src/main/java/com/fsck/k9/controller/MessageCountsProvider.kt
@@ -1,10 +1,10 @@
 package com.fsck.k9.controller
 
 import app.k9mail.legacy.account.Account
+import app.k9mail.legacy.search.ConditionsTreeNode
+import app.k9mail.legacy.search.LocalSearch
 import com.fsck.k9.mailstore.MessageStoreManager
 import com.fsck.k9.preferences.AccountManager
-import com.fsck.k9.search.ConditionsTreeNode
-import com.fsck.k9.search.LocalSearch
 import com.fsck.k9.search.SearchAccount
 import com.fsck.k9.search.excludeSpecialFolders
 import com.fsck.k9.search.getAccounts
