+++ legacy/core/src/main/java/com/fsck/k9/mailstore/MessageStore.kt
@@ -1,11 +1,11 @@
 package com.fsck.k9.mailstore
 
 import app.k9mail.legacy.account.Account.FolderMode
+import app.k9mail.legacy.search.ConditionsTreeNode
 import com.fsck.k9.mail.Flag
 import com.fsck.k9.mail.FolderClass
 import com.fsck.k9.mail.FolderType
 import com.fsck.k9.mail.Header
-import com.fsck.k9.search.ConditionsTreeNode
 import java.util.Date
 
 /**
