+++ legacy/core/src/test/java/com/fsck/k9/controller/DefaultMessageCountsProviderTest.kt
@@ -2,12 +2,12 @@ package com.fsck.k9.controller
 
 import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.account.Account.FolderMode
+import app.k9mail.legacy.search.ConditionsTreeNode
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import com.fsck.k9.Preferences
 import com.fsck.k9.mailstore.ListenableMessageStore
 import com.fsck.k9.mailstore.MessageStoreManager
-import com.fsck.k9.search.ConditionsTreeNode
 import org.junit.Test
 import org.mockito.kotlin.anyOrNull
 import org.mockito.kotlin.doReturn
