+++ legacy/storage/src/main/java/com/fsck/k9/storage/messages/K9MessageStore.kt
@@ -1,6 +1,7 @@
 package com.fsck.k9.storage.messages
 
 import app.k9mail.legacy.account.Account.FolderMode
+import app.k9mail.legacy.search.ConditionsTreeNode
 import com.fsck.k9.mail.Flag
 import com.fsck.k9.mail.FolderClass
 import com.fsck.k9.mail.FolderType
@@ -15,7 +16,6 @@ import com.fsck.k9.mailstore.MoreMessages
 import com.fsck.k9.mailstore.SaveMessageData
 import com.fsck.k9.mailstore.StorageManager
 import com.fsck.k9.message.extractors.BasicPartInfoExtractor
-import com.fsck.k9.search.ConditionsTreeNode
 import java.util.Date
 
 class K9MessageStore(
