+++ legacy/storage/src/main/java/com/fsck/k9/storage/messages/RetrieveFolderOperations.kt
@@ -4,6 +4,7 @@ import android.database.Cursor
 import androidx.core.database.getLongOrNull
 import app.k9mail.core.android.common.database.map
 import app.k9mail.legacy.account.Account.FolderMode
+import app.k9mail.legacy.search.ConditionsTreeNode
 import com.fsck.k9.mail.FolderClass
 import com.fsck.k9.mail.FolderType
 import com.fsck.k9.mailstore.FolderDetailsAccessor
@@ -12,7 +13,6 @@ import com.fsck.k9.mailstore.FolderNotFoundException
 import com.fsck.k9.mailstore.LockableDatabase
 import com.fsck.k9.mailstore.MoreMessages
 import com.fsck.k9.mailstore.toFolderType
-import com.fsck.k9.search.ConditionsTreeNode
 import com.fsck.k9.search.SqlQueryBuilder
 
 internal class RetrieveFolderOperations(private val lockableDatabase: LockableDatabase) {
