+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/messagelist/MessageListConfig.kt
@@ -1,8 +1,8 @@
 package com.fsck.k9.ui.messagelist
 
 import app.k9mail.legacy.account.Account.SortType
+import app.k9mail.legacy.search.LocalSearch
 import com.fsck.k9.controller.MessageReference
-import com.fsck.k9.search.LocalSearch
 
 data class MessageListConfig(
     val search: LocalSearch,
